requirements.txt
fastapi
uvicorn
python-dotenv
ragas
langchain
langchain-openai
langchain-anthropic
langchain-community
pydantic
datasets
numpy
pandas
sacrebleu